办公室，修改重新修改了wifi_send—_and_wait的机制,等待调试

+++ soft2/STC_LIB/USART.h
@@ -49,7 +49,7 @@ typedef struct
 	u8	B_TX_busy;		//忙标志
 
 	u16 	RX_Cnt;			//接收字节计数
-	u8	RX_TimeOut;		//接收超时
+	u16	RX_TimeOut;		//接收超时
 	u8	B_RX_EN;		//允许接收块
 	u8	B_RX_OK;		//允许接收块
